opt: Move insertBuilder to new file
Prepare to rename to mutationBuilder and generalize it for other
mutation operators.

Release note: None
